+++ src/lib.rs
@@ -8,7 +8,7 @@
 //!
 //! ```toml
 //! [dependencies]
-//! appwrite = "0.2.0"
+//! appwrite = "0.3.0"
 //! ```
 //!
 //! ## Usage
@@ -51,7 +51,7 @@ pub use input_file::InputFile;
 pub type Result<T> = std::result::Result<T, AppwriteError>;
 
 /// SDK version
-pub const VERSION: &str = "0.2.0";
+pub const VERSION: &str = "0.3.0";
 
 /// SDK name
 pub const SDK_NAME: &str = "Rust";